wip: improve webrtc folder/files structure
