+++ packages/common/src/BaseSession.ts
@@ -6,7 +6,8 @@ import Setup from './services/Setup'
 import BaseMessage from '../../common/src/messages/BaseMessage'
 import { deRegister, register, trigger, deRegisterAll } from './services/Handler'
 import BroadcastHandler from './services/BroadcastHandler'
-import { ADD, REMOVE, SwEvent, BladeMethod, NOTIFICATION_TYPE } from './util/constants'
+import { ADD, REMOVE, SwEvent, BladeMethod } from './util/constants'
+import { NOTIFICATION_TYPE } from './webrtc/constants'
 import { BroadcastParams, ISignalWireOptions, SubscribeParams, IBladeConnectResult } from './util/interfaces'
 import { Subscription, Connect, Reauthenticate, Ping } from './messages/Blade'
 import { isFunction } from './util/helpers'
