+++ packages/common/src/BrowserSession.ts
@@ -3,8 +3,8 @@ import BaseSession from './BaseSession'
 import BaseCall from './webrtc/BaseCall'
 import { ICacheDevices, IAudioSettings, IVideoSettings, BroadcastParams, SubscribeParams } from './util/interfaces'
 import { registerOnce, trigger } from './services/Handler'
-import { SwEvent, SESSION_ID, DeviceType } from './util/constants'
-import { State } from './util/constants/call'
+import { SwEvent, SESSION_ID } from './util/constants'
+import { State, DeviceType } from './webrtc/constants'
 import { getDevices, scanResolutions, removeUnsupportedConstraints, checkDeviceIdConstraints, destructSubscribeResponse, getUserMedia } from './webrtc/helpers'
 import { findElementByType } from './util/helpers'
 import { Unsubscribe, Subscribe, Broadcast } from './messages/Verto'
