+++ packages/common/src/util/interfaces.ts
@@ -66,38 +66,6 @@ export interface ISignalWireOptions {
   userVariables?: Object
 }
 
-export interface CallOptions {
-  // Required
-  destinationNumber: string
-  remoteCallerName: string
-  remoteCallerNumber: string
-  callerName: string
-  callerNumber: string
-  // Optional
-  id?: string
-  remoteSdp?: string
-  localStream?: MediaStream
-  remoteStream?: MediaStream
-  localElement?: HTMLMediaElement | string | Function
-  remoteElement?: HTMLMediaElement | string | Function
-  iceServers?: RTCIceServer[]
-  audio?: boolean | MediaTrackConstraints
-  video?: boolean | MediaTrackConstraints
-  attach?: boolean
-  useStereo?: boolean
-  micId?: string
-  micLabel?: string
-  camId?: string
-  camLabel?: string
-  speakerId?: string
-  userVariables?: Object
-  screenShare?: boolean
-  onNotification?: Function
-  googleMaxBitrate?: number
-  googleMinBitrate?: number
-  googleStartBitrate?: number
-}
-
 export interface SubscribeParams {
   channels?: string[]
   protocol?: string
