+++ packages/common/src/webrtc/BaseCall.ts
@@ -4,12 +4,12 @@ import BrowserSession from '../BrowserSession'
 import BaseMessage from '../messages/BaseMessage'
 import { Invite, Answer, Attach, Bye, Modify, Info } from '../messages/Verto'
 import Peer from './Peer'
-import { PeerType, VertoMethod, SwEvent, NOTIFICATION_TYPE, Direction } from '../util/constants'
-import { State, DEFAULT_CALL_OPTIONS, ConferenceAction, Role } from '../util/constants/call'
+import { SwEvent } from '../util/constants'
+import { State, DEFAULT_CALL_OPTIONS, ConferenceAction, Role, PeerType, VertoMethod, NOTIFICATION_TYPE, Direction } from './constants'
 import { trigger, register, deRegister } from '../services/Handler'
 import { sdpStereoHack, sdpMediaOrderHack, checkSubscribeResponse, enableAudioTracks, disableAudioTracks, toggleAudioTracks, enableVideoTracks, disableVideoTracks, toggleVideoTracks } from './helpers'
 import { objEmpty, mutateLiveArrayData, isFunction } from '../util/helpers'
-import { CallOptions } from '../util/interfaces'
+import { CallOptions } from './interfaces'
 import { attachMediaStream, detachMediaStream, sdpToJsonHack, stopStream, getUserMedia, setMediaElementSinkId } from '../util/webrtc'
 import { MCULayoutEventHandler } from './LayoutHandler'
