+++ packages/common/src/webrtc/interfaces.ts
@@ -1,0 +1,31 @@
+export interface CallOptions {
+  // Required
+  destinationNumber: string
+  remoteCallerName: string
+  remoteCallerNumber: string
+  callerName: string
+  callerNumber: string
+  // Optional
+  id?: string
+  remoteSdp?: string
+  localStream?: MediaStream
+  remoteStream?: MediaStream
+  localElement?: HTMLMediaElement | string | Function
+  remoteElement?: HTMLMediaElement | string | Function
+  iceServers?: RTCIceServer[]
+  audio?: boolean | MediaTrackConstraints
+  video?: boolean | MediaTrackConstraints
+  attach?: boolean
+  useStereo?: boolean
+  micId?: string
+  micLabel?: string
+  camId?: string
+  camLabel?: string
+  speakerId?: string
+  userVariables?: Object
+  screenShare?: boolean
+  onNotification?: Function
+  googleMaxBitrate?: number
+  googleMinBitrate?: number
+  googleStartBitrate?: number
+}
